Make the location of the initcontainer image configurable
This commit introduces the option to configure the location of the
initcontainer image in the helm chart (via the configmap), or the
deploy.yaml for skaffold. The k8s job will use the specified
imagecontainer. This gives the user the option to introduce
additional functionality before the k8s job run.