Feature/37/add working dir spec (#38)
* Refactor K8s interface to allow easier testing of `CreateK8sJob`

The `K8s` interface has been refactored, so that creation and awaiting a
k8s job are now split into two different methods. This allows the job
creation part to be tested quite easily with kubernetes fakes, as we no
longer need to fake job status updates.

Signed-off-by: Patrick Pichler <patrick.pichler@dynatrace.com>

* Add possibility to specify `workingDir` of task

It is now possible to specify the `workingDir` of a task. This means
that when the container starts, the working dir of the command
executed is set to the path specified in the task config.

Fixes #37

Signed-off-by: Patrick Pichler <patrick.pichler@dynatrace.com>