fix: separate ingress and egress network policy
In some Keptn setups there may not be a reliable way of setting up an
egress network policy by autodectecting the necessary endpoints external
to the kubernetes cluster. Since we may want to enforce the no-ingress
network policy anyway but not cripple the network communication with
external services this commit splits the two policies so that they can
be selected independently.

Signed-off-by: Paolo Chila <paolo.chila@dynatrace.com>